Changed source structure to better split content scripts and background
